% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_predict.R
\name{glm_predict}
\alias{glm_predict}
\title{GLM Prediction}
\usage{
glm_predict(model, target, threshold = 0.5, newdata = NULL, ...)
}
\arguments{
\item{model}{Fitted \code{glm} model object}

\item{target}{Target name (must be present as a column in the model object). This is the y variable when creating a \code{glm}.}

\item{threshold}{Probability threshold for predicting positive class. Must be between 0 and 1 (exclusive). default=\code{0.5}.}

\item{newdata}{New data to make predictions on. By default, \code{glm_predict} predicts on the original data used to create the model. default=\code{NULL}.}

\item{...}{Other parameters to be passed into the \code{augment} function for making predictions.}
}
\value{
A named list containing:

\verb{$augment}: Augment object containing the model predictions and information about each observation.

\verb{$accuracy}: Accuracy of predictions.

\verb{$precision}: Precision of predictions.

\verb{$recall}: Recall of predictions.
}
\description{
Create a Generalised Linear Model (glm) from a dataset and a given set of variables, then
make predictions from the model, either on the original data or a new dataset.
}
\examples{
require(dplyr)
cancer_model <- glm(malignant ~ texture_mean, data=cancer_clean, family="quasibinomial")
glm_predict(cancer_model, "malignant")
glm_predict(cancer_model, "malignant", threshold=0.8)
}
