% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{glm_predict}
\alias{glm_predict}
\title{GLM Prediction}
\usage{
glm_predict(model_data, y, x, threshold = 0.5, predict_data = NA, ...)
}
\arguments{
\item{model_data}{dataframe or tibble}

\item{y}{target variable name/response (given as a character)}

\item{x}{features/terms (given as a single character or a character vector)}

\item{threshold}{probability threshold for predicting positive class, default=0.5}

\item{predict_data}{new data to make predictions on. by default, glm_predict predicts on the original data used to create the model}

\item{...}{other parameters to be passed into glm function}
}
\value{
A named list containing:

$model: fitted model object

$augment: augment object containing the predictions and statistics

$accuracy

$precision

$recall
}
\description{
Create a Generalised Linear Model (glm) from a dataset and a given set of variables, then
make predictions from the model, either on the original data or a new dataset.
}
\examples{
require(dplyr)
glm_predict(cancer_clean, "malignant", "texture_mean")
glm_predict(cancer_clean, "malignant", c("radius_mean", "concavity_mean"), threshold=0.8)
}
