% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_score.R
\name{precision_score}
\alias{precision_score}
\title{Precision Score}
\usage{
precision_score(correct, predictions, pos_class = 1)
}
\arguments{
\item{correct}{True/correct labels.}

\item{predictions}{Predicted labels.}

\item{pos_class}{The positive prediction class.}
}
\value{
\code{precision} (numeric): The precision (true positives / true positives + false positives) of the predictions
}
\description{
Precision Score
}
