% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_score.R
\name{recall_score}
\alias{recall_score}
\title{Recall Score}
\usage{
recall_score(correct, predictions, pos_class = 1)
}
\arguments{
\item{correct}{True/correct labels.}

\item{predictions}{Predicted labels.}

\item{pos_class}{The positive prediction class.}
}
\value{
\code{recall} (numeric): The recall (true positives / true positives + false negatives) of the predictions
}
\description{
Recall Score
}
